make twitter query more flexible
