+++ data/twitterQuery.json
@@ -1,3 +1,3 @@
 {
-    "query": "((hire -job -gig -role -am) OR (\"looking for\" -job -gig -role -am) OR hiring) remote -podcast -know -how -tips -nobody -anybody -anyone -blog -filter:retweets -filter:quote"
+    "query": "(hire OR hiring OR internship OR interns OR intern OR freelance OR freelancer) -how -tips -filter:retweets -filter:quote"
 }
