+++ lib/parser.js
@@ -1,6 +1,6 @@
 const { contentSecurityPolicy } = require("helmet");
 const fetch = require("node-fetch");
-const rolesRaw = require("./data/roles.json");
+const rolesRaw = require("../data/roles.json");
 
 const categories = Object.keys(rolesRaw)
   .map((category) =>
@@ -105,16 +105,12 @@ const parseJobType = (text) => {
       if (word == "fulltime") {
         return "fulltime";
       }
-      if (word == "freshers" || word == "no experience"  || word == "fresher") {
+      if (word == "freshers" || word == "noexperience" || word == "fresher") {
          return "freshers";
       }
     }
   }
   return "fulltime";
-};
-
-const needManualVerification = (text) => {
-  return text.search(/\?/) != -1;
 };
 
 const parseRoles = (text) => {
@@ -155,7 +151,6 @@ const parseTweet = async (raw_text) => {
     type: parseJobType(raw_text),
     emails: raw_text.match(emailRegex) || [],
     urls: [...new Set(await parseURLs(raw_text))],
-    need_manual_verification: needManualVerification(raw_text),
     //stripped_text: text,
   };
 };
