--- temp.txt
@@ -1,19 +1,0 @@
-/**
-	* @swagger
-	* /api/verify:
-	*   post:
-	*     summary: Register a new user
-	*     tags:
-	*        - user
-	*     security: []
-	*     requestBody:
-	*       required: true
-	*       content:
-	*         application/json:
-	*           schema:
-	*             properties:
-	*               name:
-	*                 type: string
-	*                 description: desc
-	*                 example: abc
-	**/
